requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
docopt==0.6.2
edge-tts==7.2.7
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
num2words==0.5.14
numba==0.63.1
numpy==2.3.5
ollama==0.6.1
omegaconf==2.3.0
openai-whisper==20250625
propcache==0.4.1
pvporcupine==4.0.1
pvrecorder==1.2.7
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
setuptools==80.9.0
sounddevice==0.5.3
soundfile==0.13.1
sympy==1.14.0
tabulate==0.9.0
tiktoken==0.12.0
torch==2.9.1
torchaudio==2.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
yarl==1.22.0
